+++ src/main/java/finalmission/controller/MemberController.java
@@ -1,6 +1,12 @@
 package finalmission.controller;
 
+import finalmission.controller.config.CookieManager;
+import finalmission.controller.dto.MemberLoginRequest;
+import finalmission.controller.dto.MemberResponse;
+import finalmission.controller.dto.MemberSignupRequest;
 import finalmission.domain.Member;
+import finalmission.infrastructure.JwtTokenProvider;
+import finalmission.service.MemberService;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
