+++ src/main/java/finalmission/controller/dto/MemberLoginRequest.java
@@ -1,4 +1,4 @@
-package finalmission.controller;
+package finalmission.controller.dto;
 
 public record MemberLoginRequest(
         String email,
