+++ src/main/java/finalmission/controller/dto/MemberResponse.java
@@ -1,4 +1,4 @@
-package finalmission.controller;
+package finalmission.controller.dto;
 
 import finalmission.domain.Member;
